♻️ refactor(commands): 命令模块拆分与注册重构
将 agent 命令从单文件 commands/agent/mod.rs 拆分为独立子模块，并新增
smart_router / sub_agent / workflows 命令模块，统一通过 register_commands.rs
的 generate_handler![] 接线（符合 AGENTS.md 两步注册规则）。

- 新增：commands/smart_router.rs、commands/sub_agent/mod.rs、commands/workflows/mod.rs
- 重构：commands/agent/mod.rs 大幅精简（-741 行），命令职责下沉到子模块
- 扩展：evaluator / evolution / memory / user_profile / agent_analytics 命令
- crate 修复：disk-cache、search(ast_index/file_index)、mcp、tools/knowledge、agent evaluator/runner
- 删除临时文件 fix_encoding.txt；新增 REFACTOR_PLAN.md 记录重构方案
- 新增命令已在 lib.rs(mod 声明) + commands/mod.rs(pub mod) + register_commands.rs(generate_handler) 三处完整注册

dprint check 通过。

+++ src-tauri/crates/tools/src/tools/knowledge.rs
@@ -177,7 +177,7 @@ impl Tool for SearchKnowledgeTool {
         })?;
 
         let rows: Vec<String> = stmt
-            .query_map(rusqlite::params![like_pattern, top_k], |row| {
+            .query_map(rusqlite::params![like_pattern, top_k as i64], |row| {
                 let content: String = row.get(0)?;
                 Ok(content)
             })
